refactor: Move dark theme markup
- Move the ui and syntax related colours into colour palette objects.
- Move existing theme-related stuff into a separate file.
